If SELF is undefined, return null instead
If both /software/components/profile/env and VO_ATLAS_LOCAL_AREA are undefined,
the result will be that SELF is undefined, settings /software/components/profile/env
to undefined will cause a really ugly validation error, so instead we will return null
which is perfectly valid in this case.

Thanks to @jouvin for the "proper" fix.