support for muti options settings for logical volumes
DISK_LV_BY_VG is updated to support setting of logical volume setting, such as type.
which sets the raid_type if set.
using the new function lvo_add which supports taking lv options as dict to set size and type.

Creating raid through LVM is now possible but there appears to be no pan settings to set the logical volume type.
this will provide support for this.

fixes #150